+++ default-notice.php
@@ -1,4 +1,4 @@
-<?php // Copyright (c) 2017, SWITCH ?>
+<?php // Copyright (c) 2018, SWITCH ?>
 
 <!-- Identity Provider Selection: Start -->
 <h1><?php echo getLocalString('settings'); ?></h1>